+++ src/main/java/com/bud/feature/bud/reaction/BudReactionEntry.java
@@ -5,10 +5,6 @@ import javax.annotation.Nonnull;
 import com.bud.core.components.BudComponent;
 import com.bud.feature.queue.IQueueEntry;
 
-/**
- * One bud reacting in character to something that happened to another bud
- * (mood change, a new bud spawning, a legendary memory being formed).
- */
 public record BudReactionEntry(
         @Nonnull BudComponent budComponent,
         @Nonnull BudReactionKind kind,